Some stuff for testing in here now, like objects that spin based on user vel and pos and what not.
But also FIXED problem where sometimes user could jump and when he hits the ground they would still be in the falling position.

Still working on the weird jumping thing where you can't jump and move to the right as fast as you can to the left...probablt something to do with the collision logic?